fix: extract tool constants to resolve circular dependency
