+++ src/swagger/client.ts
@@ -397,10 +397,14 @@ export class SwaggerClient implements Client {
 
           const result = await handlerFn.call(this, args);
 
-          // Use custom formatter if available, otherwise return JSON
           const formattedResult = formatResponse
             ? formatResponse(result)
             : result;
+
+          if (toolParams.outputSchema) {
+            return { structuredContent: formattedResult, content: [] };
+          }
+
           const responseText =
             typeof formattedResult === "string"
               ? formattedResult
@@ -411,6 +415,7 @@ export class SwaggerClient implements Client {
           };
         } catch (error) {
           return {
+            isError: true,
             content: [
               {
                 type: "text",
