+++ src/swagger/client/tools.ts
@@ -8,6 +8,7 @@
 
 import type { ToolParams } from "../../common/types";
 import { FUNCTIONAL_TESTING_TOOLS } from "./functional-testing-tools";
+import { DELETING, MUTATING, READ_ONLY } from "./tool-constants";
 import {
   CreatePortalArgsSchema,
   CreateProductArgsSchema,
@@ -55,25 +56,6 @@ export interface SwaggerToolParams extends ToolParams {
   handler: string;
   formatResponse?: (result: any) => any;
 }
-
-export const READ_ONLY = {
-  readOnly: true,
-  destructive: false,
-  openWorld: false,
-} as const;
-
-const MUTATING = {
-  readOnly: false,
-  destructive: false,
-  openWorld: true,
-} as const;
-
-const DELETING = {
-  readOnly: false,
-  destructive: true,
-  openWorld: true,
-} as const;
-
 export const TOOLS: SwaggerToolParams[] = [
   {
     title: "List Portals",
